move custom handlers to tp_conv.py
